feat(server): 添加系统菜单管理功能
- 新增系统菜单模型、请求、响应及相关服务
- 添加菜单路由和API接口
- 更新用户服务以支持菜单关联
- 调整用户ID类型为int64以统一数据类型
- 新增自定义校验规则用于菜单管理

+++ server/model/model.go
@@ -7,7 +7,7 @@ import (
 )
 
 type IdWrapper struct {
-	Id int64 `json:"id" gorm:"primarykey;comment:主键"`
+	Id uint `json:"id" gorm:"primarykey;comment:主键"`
 }
 
 type ControlWrapper struct {
